requirements.txt
Flask
gunicorn
python-dotenv
requests
urllib3
Werkzeug
